fixes #106

+++ app/source/classes/Models/Plugin.php
@@ -126,6 +126,10 @@ class Plugin implements ModelInterface {
     }
     
     public static function create($plugin){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         // Is the name valid?
         if(!mb_strlen($plugin['name']) || Leafpub::isProtectedSlug($plugin['name'])) {
             throw new \Exception('Invalid name: ' . $plugin['name'], self::INVALID_NAME);
@@ -162,6 +166,10 @@ class Plugin implements ModelInterface {
     }
     
     public static function edit($plugin){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         // Is the name valid?
         if(!mb_strlen($plugin['name']) || Leafpub::isProtectedSlug($plugin['name'])) {
             throw new \Exception('Invalid name: ' . $plugin['name'], self::INVALID_NAME);
@@ -198,6 +206,10 @@ class Plugin implements ModelInterface {
     }
     
     public static function delete($plugin){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         try {
            $rowCount = self::getModel()->delete(['dir' => $plugin]);
            return ($rowCount > 0);
@@ -267,6 +279,10 @@ class Plugin implements ModelInterface {
     *
     */
     public static function install($zipFile){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         $bPluginJson = false;
         $bPluginPhp = false;
         $bUpdate = false;
@@ -342,6 +358,10 @@ class Plugin implements ModelInterface {
     *
     */
     public static function deinstall($plugin){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         if (!is_dir(Leafpub::path("content/plugins/$plugin"))){
             return false;
         }
@@ -370,6 +390,10 @@ class Plugin implements ModelInterface {
     *
     */
     public static function activate($dir){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+        
         // Always sanitize user input ;-)
         if (!is_dir(Leafpub::path("content/plugins/$dir"))){
             return false;
@@ -423,6 +447,10 @@ class Plugin implements ModelInterface {
     *
     */
     public static function deactivate($dir){
+        if (!self::isAllowedCaller()){
+            return false;
+        }
+
         if (!is_dir(Leafpub::path("content/plugins/$dir"))){
             return false;
         }
